requirements.txt
altair==5.2.0
anyio==4.1.0
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==5.5
decorator==4.4.2
deprecation==2.1.0
executing==2.0.1
eyed3==0.9.7
filetype==1.2.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mps-youtube==0.2.8
mutagen==1.47.0
numpy==1.26.2
packaging==23.2
pafy==0.5.5
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==4.25.1
pure-eval==0.2.2
pyarrow==14.0.1
pycryptodomex==3.19.1
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
simpleaudio==1.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.28.2
streamlit-autorefresh==1.0.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.12
websockets==12.0
youtube-dl==2021.12.17
youtube-search==2.1.2
youtube-search-python==1.6.6
yt-dlp==2023.12.30
ytmusicapi==1.4.1
zipp==3.17.0
websockets
asyncio
base64
json
pyaudio
os
pathlib